social_network/requirements.txt
django
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
django-filter
django-redis
